refactor: split ClassUtils into cache resolvers
