+++ fesod-sheet/src/main/java/org/apache/fesod/sheet/util/SheetContentPropertyResolver.java
@@ -1,0 +1,214 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.fesod.sheet.util;
+
+import java.lang.reflect.Field;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import org.apache.fesod.common.util.MapUtils;
+import org.apache.fesod.shaded.cglib.beans.BeanMap;
+import org.apache.fesod.sheet.annotation.ExcelProperty;
+import org.apache.fesod.sheet.annotation.format.DateTimeFormat;
+import org.apache.fesod.sheet.annotation.format.NumberFormat;
+import org.apache.fesod.sheet.annotation.write.style.ContentFontStyle;
+import org.apache.fesod.sheet.annotation.write.style.ContentStyle;
+import org.apache.fesod.sheet.converters.AutoConverter;
+import org.apache.fesod.sheet.converters.Converter;
+import org.apache.fesod.sheet.enums.CacheLocationEnum;
+import org.apache.fesod.sheet.exception.ExcelCommonException;
+import org.apache.fesod.sheet.metadata.ConfigurationHolder;
+import org.apache.fesod.sheet.metadata.property.DateTimeFormatProperty;
+import org.apache.fesod.sheet.metadata.property.ExcelContentProperty;
+import org.apache.fesod.sheet.metadata.property.FontProperty;
+import org.apache.fesod.sheet.metadata.property.NumberFormatProperty;
+import org.apache.fesod.sheet.metadata.property.StyleProperty;
+import org.apache.fesod.sheet.util.ClassUtils.ContentPropertyKey;
+
+/**
+ * Resolves the converter, format and style of a single field, merged from the head class and the
+ * runtime class of the data.
+ */
+final class SheetContentPropertyResolver {
+
+    /**
+     * The cache configuration information for each of the class
+     */
+    private static final ThreadLocal<Map<Class<?>, Map<String, ExcelContentProperty>>> CLASS_CONTENT_THREAD_LOCAL =
+            new ThreadLocal<>();
+
+    /**
+     * The cache configuration information for each of the class
+     */
+    private static final ThreadLocal<Map<ContentPropertyKey, ExcelContentProperty>> CONTENT_THREAD_LOCAL =
+            new ThreadLocal<>();
+
+    private static final Map<CacheLocationEnum, MetadataCacheStrategy<ContentPropertyKey, ExcelContentProperty>>
+            CONTENT_STRATEGIES = MetadataCacheStrategy.byLocation(ClassUtils.CONTENT_CACHE, CONTENT_THREAD_LOCAL);
+
+    private static final Map<CacheLocationEnum, MetadataCacheStrategy<Class<?>, Map<String, ExcelContentProperty>>>
+            CLASS_CONTENT_STRATEGIES =
+                    MetadataCacheStrategy.byLocation(ClassUtils.CLASS_CONTENT_CACHE, CLASS_CONTENT_THREAD_LOCAL);
+
+    private SheetContentPropertyResolver() {}
+
+    /**
+     * Calculate the configuration information for the class
+     *
+     * @param dataMap
+     * @param headClazz
+     * @param fieldName
+     * @return
+     */
+    static ExcelContentProperty resolve(
+            Map<?, ?> dataMap, Class<?> headClazz, String fieldName, ConfigurationHolder configurationHolder) {
+        Class<?> clazz = null;
+        if (dataMap instanceof BeanMap) {
+            Object bean = ((BeanMap) dataMap).getBean();
+            if (bean != null) {
+                clazz = bean.getClass();
+            }
+        }
+        return getExcelContentProperty(clazz, headClazz, fieldName, configurationHolder);
+    }
+
+    static void removeThreadLocalCache() {
+        CLASS_CONTENT_THREAD_LOCAL.remove();
+        CONTENT_THREAD_LOCAL.remove();
+    }
+
+    private static ExcelContentProperty getExcelContentProperty(
+            Class<?> clazz, Class<?> headClass, String fieldName, ConfigurationHolder configurationHolder) {
+        return MetadataCacheStrategy.select(
+                        CONTENT_STRATEGIES,
+                        configurationHolder.globalConfiguration().getFiledCacheLocation())
+                .get(
+                        buildKey(clazz, headClass, fieldName),
+                        key -> doGetExcelContentProperty(clazz, headClass, fieldName, configurationHolder));
+    }
+
+    private static ExcelContentProperty doGetExcelContentProperty(
+            Class<?> clazz, Class<?> headClass, String fieldName, ConfigurationHolder configurationHolder) {
+        ExcelContentProperty excelContentProperty = Optional.ofNullable(
+                        declaredFieldContentMap(clazz, configurationHolder))
+                .map(map -> map.get(fieldName))
+                .orElse(null);
+        ExcelContentProperty headExcelContentProperty = Optional.ofNullable(
+                        declaredFieldContentMap(headClass, configurationHolder))
+                .map(map -> map.get(fieldName))
+                .orElse(null);
+        ExcelContentProperty combineExcelContentProperty = new ExcelContentProperty();
+
+        combineExcelContentProperty(combineExcelContentProperty, headExcelContentProperty);
+        if (clazz != headClass) {
+            combineExcelContentProperty(combineExcelContentProperty, excelContentProperty);
+        }
+        return combineExcelContentProperty;
+    }
+
+    static void combineExcelContentProperty(
+            ExcelContentProperty combineExcelContentProperty, ExcelContentProperty excelContentProperty) {
+        if (excelContentProperty == null) {
+            return;
+        }
+        if (excelContentProperty.getField() != null) {
+            combineExcelContentProperty.setField(excelContentProperty.getField());
+        }
+        if (excelContentProperty.getConverter() != null) {
+            combineExcelContentProperty.setConverter(excelContentProperty.getConverter());
+        }
+        if (excelContentProperty.getDateTimeFormatProperty() != null) {
+            combineExcelContentProperty.setDateTimeFormatProperty(excelContentProperty.getDateTimeFormatProperty());
+        }
+        if (excelContentProperty.getNumberFormatProperty() != null) {
+            combineExcelContentProperty.setNumberFormatProperty(excelContentProperty.getNumberFormatProperty());
+        }
+        if (excelContentProperty.getContentStyleProperty() != null) {
+            combineExcelContentProperty.setContentStyleProperty(excelContentProperty.getContentStyleProperty());
+        }
+        if (excelContentProperty.getContentFontProperty() != null) {
+            combineExcelContentProperty.setContentFontProperty(excelContentProperty.getContentFontProperty());
+        }
+    }
+
+    private static ContentPropertyKey buildKey(Class<?> clazz, Class<?> headClass, String fieldName) {
+        return new ContentPropertyKey(clazz, headClass, fieldName);
+    }
+
+    private static Map<String, ExcelContentProperty> declaredFieldContentMap(
+            Class<?> clazz, ConfigurationHolder configurationHolder) {
+        if (clazz == null) {
+            return null;
+        }
+        return MetadataCacheStrategy.select(
+                        CLASS_CONTENT_STRATEGIES,
+                        configurationHolder.globalConfiguration().getFiledCacheLocation())
+                .get(clazz, key -> doDeclaredFieldContentMap(clazz));
+    }
+
+    private static Map<String, ExcelContentProperty> doDeclaredFieldContentMap(Class<?> clazz) {
+        if (clazz == null) {
+            return null;
+        }
+        List<Field> tempFieldList = FieldUtils.resolveAllFields(clazz);
+
+        ContentStyle parentContentStyle = clazz.getAnnotation(ContentStyle.class);
+        ContentFontStyle parentContentFontStyle = clazz.getAnnotation(ContentFontStyle.class);
+        Map<String, ExcelContentProperty> fieldContentMap = MapUtils.newHashMapWithExpectedSize(tempFieldList.size());
+        for (Field field : tempFieldList) {
+            ExcelContentProperty excelContentProperty = new ExcelContentProperty();
+            excelContentProperty.setField(field);
+
+            ExcelProperty excelProperty = field.getAnnotation(ExcelProperty.class);
+            if (excelProperty != null) {
+                Class<? extends Converter<?>> convertClazz = excelProperty.converter();
+                if (convertClazz != AutoConverter.class) {
+                    try {
+                        Converter<?> converter =
+                                convertClazz.getDeclaredConstructor().newInstance();
+                        excelContentProperty.setConverter(converter);
+                    } catch (Exception e) {
+                        throw new ExcelCommonException("Can not instance custom converter:" + convertClazz.getName());
+                    }
+                }
+            }
+
+            ContentStyle contentStyle = field.getAnnotation(ContentStyle.class);
+            if (contentStyle == null) {
+                contentStyle = parentContentStyle;
+            }
+            excelContentProperty.setContentStyleProperty(StyleProperty.build(contentStyle));
+
+            ContentFontStyle contentFontStyle = field.getAnnotation(ContentFontStyle.class);
+            if (contentFontStyle == null) {
+                contentFontStyle = parentContentFontStyle;
+            }
+            excelContentProperty.setContentFontProperty(FontProperty.build(contentFontStyle));
+
+            excelContentProperty.setDateTimeFormatProperty(
+                    DateTimeFormatProperty.build(field.getAnnotation(DateTimeFormat.class)));
+            excelContentProperty.setNumberFormatProperty(
+                    NumberFormatProperty.build(field.getAnnotation(NumberFormat.class)));
+
+            fieldContentMap.put(field.getName(), excelContentProperty);
+        }
+        return fieldContentMap;
+    }
+}
